Optimization: CreateAccount only clears state if code hash exists (#2255)
* only clear state if code size key exists

* use hash instead

+++ x/evm/state/state.go
@@ -146,11 +146,12 @@ func (s *DBImpl) clearAccountStateIfDestructed(st *TemporaryState) {
 
 func (s *DBImpl) clearAccountState(acc common.Address) {
 	s.k.PrepareReplayedAddr(s.ctx, acc)
-	s.k.PurgePrefix(s.ctx, types.StateKey(acc))
-	deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeKeyPrefix), acc[:])
-	deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeSizeKeyPrefix), acc[:])
-	deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeHashKeyPrefix), acc[:])
-	deleteIfExists(s.k.PrefixStore(s.ctx, types.NonceKeyPrefix), acc[:])
+	if deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeHashKeyPrefix), acc[:]) {
+		s.k.PurgePrefix(s.ctx, types.StateKey(acc))
+		deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeKeyPrefix), acc[:])
+		deleteIfExists(s.k.PrefixStore(s.ctx, types.CodeSizeKeyPrefix), acc[:])
+		deleteIfExists(s.k.PrefixStore(s.ctx, types.NonceKeyPrefix), acc[:])
+	}
 }
 
 func (s *DBImpl) MarkAccount(acc common.Address, status []byte) {
@@ -204,8 +205,10 @@ func (s *DBImpl) getTransientState(acc common.Address, key common.Hash) (common.
 	return val, found
 }
 
-func deleteIfExists(store storetypes.KVStore, key []byte) {
+func deleteIfExists(store storetypes.KVStore, key []byte) bool {
 	if store.Has(key) {
 		store.Delete(key)
+		return true
 	}
+	return false
 }
